@@ -7,6 +7,7 @@
     "type": "git",
     "url": "https://github.com/burnnat/webscreens.git"
   },
+  "type": "module",
   "files": [
     "dist",
     "public",
@@ -22,44 +23,43 @@
     "postversion": "git push && git push --tags"
   },
   "dependencies": {
-    "body-parser": "^1.13.3",
-    "cookie-parser": "~1.3.5",
+    "body-parser": "^1.20.2",
+    "cookie-parser": "~1.4.6",
     "exif": "^0.6.0",
-    "express": "^4.13.4",
+    "express": "^4.18.2",
     "gaze": "^1.1.3",
-    "glob": "^5.0.14",
-    "lodash": "^4.17.15",
-    "moment-timezone": "^0.5.26",
-    "morgan": "^1.6.1",
+    "glob": "^10.3.5",
+    "lodash": "^4.17.21",
+    "moment-timezone": "^0.5.43",
+    "morgan": "^1.10.0",
     "moveto": "^1.8.2",
-    "mustache-express": "^1.2.8",
-    "mythtv-services-api": "^1.1.3",
-    "our-groceries-client": "^0.1.65-pre",
-    "puppeteer": "^2.0.0",
+    "mustache-express": "^1.3.2",
+    "mythtv-services-api": "^31.0.6",
+    "nanoid": "^5.0.1",
+    "our-groceries-client": "0.1.65-pre",
+    "puppeteer": "^21.3.1",
     "serve-favicon": "^2.5.0",
-    "sharp": "^0.22.1",
-    "shortid": "^2.2.14",
+    "sharp": "^0.32.6",
     "text-to-svg": "^3.1.5",
-    "yargs": "^13.2.4"
+    "yargs": "^17.7.2"
   },
   "devDependencies": {
-    "@types/chai": "^4.1.3",
-    "@types/cookie-parser": "^1.4.1",
-    "@types/exif": "^0.6.0",
-    "@types/express": "^4.0.36",
-    "@types/lodash": "^4.14.68",
-    "@types/mocha": "^5.2.0",
-    "@types/node": "^8.0.9",
-    "@types/puppeteer": "^1.19.1",
-    "@types/sinon": "^4.3.3",
-    "chai": "^4.1.2",
-    "concurrently": "^3.5.0",
-    "fs-extra": "^8.1.0",
-    "mocha": "^5.1.1",
-    "node-mocks-http": "^1.7.0",
-    "nodemon": "^1.17.4",
-    "sinon": "^5.0.7",
-    "ts-node": "^6.0.3",
-    "typescript": "^2.4.1"
+    "@tsconfig/node18": "^18.2.2",
+    "@types/chai": "^4.3.6",
+    "@types/cookie-parser": "^1.4.4",
+    "@types/exif": "^0.6.3",
+    "@types/express": "^4.17.17",
+    "@types/lodash": "^4.14.198",
+    "@types/mocha": "^10.0.1",
+    "@types/node": "^18.1.9",
+    "chai": "^4.3.8",
+    "concurrently": "^8.2.1",
+    "fs-extra": "^11.1.1",
+    "mocha": "^10.2.0",
+    "node-mocks-http": "^1.13.0",
+    "nodemon": "^3.0.1",
+    "sinon": "^16.0.0",
+    "ts-node": "^10.9.1",
+    "typescript": "^5.2.2"
   }
 }
