@@ -1,2 +1,3 @@
 #!/usr/bin/env node
-require('./dist/index.js').start();
+import { start } from './dist/index.js';
+start();
