@@ -1,9 +1,9 @@
-import * as sinon from 'sinon';
+import sinon from 'sinon';
 import { expect, assert } from "chai";
-import * as mocks from 'node-mocks-http';
+import mocks from 'node-mocks-http';
 import { Example } from '../src/models/example.model';
 import { after } from 'mocha';
-import config from "../src/config/config";
+import config from "../src/config/config.js";
 var mongoose = require('mongoose');
 
 describe("Model Spec", function () {
